fix bug on kloosterman

+++ DirichletCharacter.py
@@ -308,8 +308,8 @@ def dc_calc_kloosterman(modulus,number):
         chi = DirichletGroup_conrey(modulus)[number]
         chi = chi.sage_character()
         k = chi.kloosterman_sum_numerical(100,arg[0],arg[1])
-        real = int(round(k.real(),5))
-        imag = int(round(k.imag(),5))
+        real = round(k.real(),5)
+        imag = round(k.imag(),5)
         if imag == 0:
             k = str(real)
         elif real == 0:
